Moved the bracket form into a component
Having issues on variable definition warnings in the console
